Per code review, separate in a different file #delete, #delete_if and #grep_v
